requirements.txt
openai
python-dotenv
openlimit
faker
pandas
openpyxl
jinja2
